control_code/requirements.txt
aioconsole==0.1.16
apriltag==0.0.16
astroid==2.4.1
begins==0.9
certifi==2020.4.5.2
cycler==0.10.0
Cython==0.29.17
easycython==1.0.7
flake8==3.8.2
future==0.18.2
imutils==0.5.3
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.1.3
mccabe==0.6.1
numpy==1.18.1
opencv-contrib-python==4.2.0.34
pandas==1.0.3
pycodestyle==2.6.0
pycpd==2.0.0
pyflakes==2.2.0
pylint==2.5.2
pyparsing==2.4.7
pyserial==3.4
pyserial-asyncio==0.4
python-dateutil==2.8.1
pytz==2020.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
toml==0.10.0
wrapt==1.11.2
